Split main off
